pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-orb"
version = "0.1.0"
description = "Custom intraday futures research and backtesting framework with ORB-first architecture."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Prana" }
]
dependencies = [
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "pydantic>=2.7,<3.0",
  "pyarrow>=16.0,<21.0",
  "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.4,<1.0",
  "mypy>=1.10,<2.0",
]
ml = [
  "scikit-learn>=1.5,<2.0",
]
qlib = [
  "pyqlib>=0.9,<1.0",
]
finrl = [
  "finrl>=0.3,<1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
